requirements.txt
Django==5.0
djangorestframework==3.14.0
drf_yasg==1.21.7
pyotp==2.9.0
rest_framework_simplejwt==0.0.2
djangorestframework-simplejwt
django_redis
psycopg2-binary==2.9.9